+++ src/modules/dashboard/components/ActionsPageFeed/index.ts
@@ -1,1 +1,3 @@
 export { default } from './ActionsPageFeed';
+export { default as ActionsPageEvent } from './ActionsPageEvent';
+export { default as ActionsPageFeedItem } from './ActionsPageFeedItem';
